CommandParserText

Description

W module znajdują się funkcje związane z parsowaniem tekstu od użytkownika. Jest tu deklaracja typu danych Action, który reprezentuje obsługiwane przez program komendy. Dane wpisywana przez użytkownika są parsowane do odpowiednich akcji. Moduł zawiera również obsługiwane synonimy bazowych komend. Dane są przetwarzane w formacie Text.

Documentation

data Action

Typ danych reprezenujący akcje obsługiwane przez program

Constructors

Move Direction
akcja przemieszczenia się między pokojami

Take ItemName
akcja zabrania przedmiotu

Put ItemName
akcja odłożenia przedmiotu

Help
wypisanie na konsolę informacji pomocy

Restart
funkcja restartująca stan gry

SaveState String
zapisanie aktualnego stanu gry do podanego pliku

Continue String
wczytanie zapisanego stanu gry z podanego pliku

Quit
wyjście z programu

Inventory
wypisanie posiadanych przedmiotów

Look
wypisanie opisu pokoju w którym gracz się znajduje

Use ItemName UserActionInput
użycie przedmiotu o danej nazwie w podany sposób

goSynonyms :: [Text]

Akceptowalne synonimy komendy "go".

takeSynonyms :: [Text]

Akceptowalne synonimy komendy "take".

putSynonyms :: [Text]

Akceptowalne synonimy komendy "put".

inventorySynonyms :: [Text]

Akceptowalne synonimy komendy "inventory".

findAction :: [Text] -> Maybe Action

Funkcja findAction przyjmuje tablicę wyrazów i próbuje dopasować pierwszy wyraz do odpowiedniej komendy (oprócz komendy Use). Jeśli nastąpi dopasowanie zwracana jest opakowana odpowiednia akcja, w razie potrzeby z jej argumentem. Jeśli nie zajdzie dopasowanie zwracane jest Nothing.

wordParser :: Parser Text

Funkcja wordParser parsuje kolejne litery słowa i zamiena je na lowercase. Funckja korzysta z parsera importowanego z modułu Text.

inputParser :: Parser [Text]

Funkcja parsuje wejście rozdzielając dane na podstawie spacji, aby uzyskać osobne słowa z polecenia. Każde słowo jest parsowane przez funckję wordParser.

commandParser :: Parser Action

Funkcja parsuje wejście za pomocą inputParser oraz próbuje znaleźć odpowiednią akcję dzięki funkcji findAction.

parseActionFromMap :: Text -> ActionsMap -> Maybe Action

Funkcja parsuje wejście poprzez próbę dopasowania go do komend znajdujących się w zbiorze typu ActionsMap. Jeśli dopasowanie się powiedzie zwracana jest opakowana akcja Use z odpowiednimi argumentami. W przeciwnym przypadku zwracane jest Nothing.

parseCommand :: ActionsMap -> Text -> Either ParseError Action

Funkcja przyjmuje zbiór akcji typu ActionsMap i wejście od użytkownika. Wejście jest parsowane z użyciem funkcji commandParser. Jeśli nie zostanie prawidłowo sparsowane to wejście może zostać dopasowane do akcji ze zbioru typu ActionsMap. Funcja zwraca Right Action jeśli udało się dopasować akcję, lub Left String jeśli wystąpił błąd lub nie nastąpiło żadne dopasowanie.

parseDirection :: Text -> Maybe Direction

Funkcja przyjmuje dane Text i próbuje dopasować je do typu Direction. Zwracane jest Just Direction lub Nothing, jeśli nie nastąpiło przypisanie.