GameActions

Description

Ten moduł zawiera definicje funkcji obsługującyh akcje związane z wykorzystaniem przedmiotów. Dany przedmiot może być użyty tylko jeśli została dla niego zdefiniowana akcja typu GameAction o nazwie ActionName.

Documentation

type ActionsMap = Map ActionName GameAction

Typ do zarządzania akcjami dla przedmiotów. Jest to mapa, której kluczami są typy ActionName a wartościami 'GameAction.

type ActionName = Text

Alias dla nazwy akcji.

type ActionMessage = String

Alias dla wiadomości w przypadku sukcesu akcji.

type UserActionInput = Text

Alias dla wejścia od użytkownika.

data ActionType

Typ danych reprezentujący rodzaje akcji związanych z przedmiotami.

Constructors

AddItem ItemName
dodanie przedmiotu

RemoveItem ItemName
usunięcie przedmiotu

ReplaceItem ItemName ItemName
zamienienie przedmiotu

data ActionResult

Typ danych reprezentujący rodzaje wyników akcji związanych z przedmiotami.

Constructors

ChangeCurrentRoom RoomName
zmiana aktualnego pokoju na inny

ChangeInventory ActionType
zmiana posiadanych przedmiotów

ChangeRoomInventory ActionType
zmiana przedmiotów znajdującyh się w aktualnym pokoju

DoNothing
brak zmiany świata gry

data ActionConditions

Typ danych reprezentujący warunki pomyślnego wykonania akcji.

Constructors

ActionConditions

Fields

roomName :: RoomName

nazwa pokoju w którym akcja musi się odbyć

itemsInRoom :: [ItemName]

lista przedmiotów znajdujących się w aktualnym pokoju

data GameAction

Typ danych reprezentujący akcję związaną z przedmiotem.

Constructors

GameAction

Fields

itemName :: ItemName

nazwa przedmiotu którego dotyczy akcja

message :: ActionMessage

wiadomość wyświetlania po pozytywnym wykonaniu akcji

conditions :: ActionConditions

warunki niezbędne do pozytywnego wykonania akcji

actionResult :: ActionResult

wynik działania akcji

executeActionResult :: ActionResult -> AdventureGame -> AdventureGame

Funkcja przyjmuje wynik akcji ActionResult i w zależności od jego typu wywłuje odpowiednie funckje zmieniające stan gry.

doesRoomContainItems :: Room -> [ItemName] -> Bool

Funkcja przyjmuje pokój i listę przedmiotów. Zwraca True jeśli wszystkie przedmioty z listy znajdują się w pokoju. Wpp zwraca False.

checkConditions :: ActionConditions -> AdventureGame -> Bool

Funkcja przyjmuje warunki typu ActionConditions i sprawdza, czy wszystkie warunki do wykonania akcji są spełnione.

canActionBeExecuted :: UserActionInput -> ItemName -> ActionsMap -> AdventureGame -> Bool

Funkcja przyjmuje dane wejściowe od użytkownika, nazwę przedmiotu. Sprawdza czy akcja może zostać wykonana, to znaczy czy dany przedmiot jest w posiadaniu, czy istnieje akcja o danej nazwie i czy wszystkie potrzebne do niej warunki są spełnione. Zwraca True lub False.

getActionResult :: ActionName -> ActionsMap -> GameAction

Funkcja wyciąga z mapy GameAction dla podanej nazwy akcji. Uwaga: tą funkcę można wywoływać tylko po sprawdzeniu, że w mapie znajduje się akcja o podanej nazwie.

runAction :: ActionName -> ActionsMap -> GameIO ()

Funkcja wykonuje daną akcję poprzez modyfikację stanu gry. Funkcja wypisuje stosowną wiadomość o sukcesie lub porażce.

useItem :: UserActionInput -> ItemName -> ActionsMap -> GameIO Bool

Funkcja przyjmuje wejście od użytkownika, nazwę przedmiotu i mapę akcji. Akcja zostaje sprawdzona funkcją canActionBeExecuted i ewentualnie wykonana funkcją runAction.

loadMapFromFile :: IO (Either String (Map ActionName GameAction))

Funkca parsuje flik z akcjami. Plik powinien się znajodwać w folderze gameFiles i mieć nazwę actions.

parseActions :: IO ActionsMap

Funkcja próbuje sparsować plik z akcjami. Zwraca sparsowaną mapę akcji lub wypisuje wiadomość o porażce i zwraca pustą mapę.

checkInputActionsMap :: UserActionInput -> ActionsMap -> Bool

Funkcja sprawdza czy dana akcja z wejścia użytkownika znajduje się w zbiorze wszystkich akcji.