GameState

Description

Ten moduł zawiera definicje typu stanu gry oraz funkcji z tym związanych. Zawiera również funkcje odpowiadające za zapis i odczyt stanu gry z pliku.

Documentation

type GameIO a = StateT AdventureGame IO a

Alias dla skrócenia zapisu.

data AdventureGame

Typ reprezentujący stan gry

Constructors

AdventureGame

Fields

gameWorld :: World

mapa zawierające wszystkie pokoje, czyli świat gry

inventory :: [ItemName]

lista zawierająca posiadane przedmioty

currentRoomName :: RoomName

nazwa pokoju w którym aktualnie znajduje się gracz

type World = Map RoomName Room

Alias dla mapy pokojów, czyli świata gry.

saveStateToFile :: FilePath -> AdventureGame -> IO ()

Funkcja koduje i zapisuje stan gry typu AdventureGame do podanego pliku. Plik zostaje zapisany do folderu gameFiles.

loadStateFromFile :: FilePath -> IO (Either String AdventureGame)

Funkcja odczytuje stan gry typu AdventureGame z podanego pliku. Plik powinien się znajdować w folderze gameFiles.

loadState :: FilePath -> GameIO ()

Funkcja próbuje załadować stan gry z podanego pliku za pomocą funkcji loadStateFromFile. Jeśli stan gry zostanie poprawnie sparsowany, jest on ustawiany jako aktualny. Wpp wypisywany jest błąd.

saveState :: FilePath -> GameIO ()

Funkcja zapisuje stan gry za pomocą funkcji saveStateToFile.

getCurrentRoom :: AdventureGame -> Room

Funkcja zwracająca aktualny pokój. Jeśli pokój nie zostanie odnaleziony poprzez nazwę, zwracany jest "zgubiony pokój".

getCurrentRoom' :: AdventureGame -> Maybe Room

Funkcja zwracająca aktualny pokój w monadzie Maybe. Jeśli pokój nie zostanie odnaleziony poprzez nazwę, zwracane jest Nothing

restartGame :: GameIO Bool

Funkcja restartuje grę ustawiając nowy stan gry.

showGameFile :: FilePath -> IO ()

Funkcja wyświetla zawartość podanego pliku znajdującego się w folderze gameFiles.

showHelp :: GameIO Bool

Funkcja wyświetla plik zawierający instrukcję pomocy dla gracza. Plik powinien się znajdować w folderze gameFiles i mieć nazwę help.txt.

showWelcome :: IO ()

Funkcja wyświetla plik zawierający powitanie dla gracza. Plik powinien się znajdować w folderze gameFiles i mieć nazwę welcome.txt.

showFirstRoom :: RoomName -> World -> IO ()

Funkcja wyświetla pokój o podanej nazwie. Jest używana do wyświetlania opisu startowego pokoju.

processActionWithMessage :: String -> GameIO ()

Funkcja processActionWithMessage służy do wypisywania podanej wiadomości.

acceptAction :: GameIO ()

Funkcja acceptAction służy do wypisywania wiadomości o powodzeniu wykonania akcji.

rejectAction :: GameIO ()

Funkcja rejectAction służy do wypisywania wiadomości o niepowodzeniu wykonania akcji.