Items

Description

Ten moduł zawiera definicje funkcji związanych z podnoszeniem i odkładaniem przedmiotów w pokoju. Znajdują się tu funkcje odpowiadające zarówno za zmianę pokoi jak i zmianę posiadanych przedmiotów.

Documentation

addItemToInventory :: ItemName -> AdventureGame -> AdventureGame

Funkcja przyjmuje nazwę przedmiotu i dodaje przedmiot do inwentarza.

removeItemFromInventory :: ItemName -> AdventureGame -> AdventureGame

Funkcja przyjmuje nazwę przedmiotu i usuwa przedmiot do inwentarza.

itemPossesion :: ItemName -> AdventureGame -> Bool

Funkcja przyjmuje nazwę przedmiotu i sprawdza czy znajduje się on w inwentarzu.

lookForItemInRoom :: ItemName -> AdventureGame -> Maybe Bool

Funkcja przyjmuje nazwę przedmiotu i sprawdza czy znajduje się on w aktualnym pokoju gracza.

takeItemChangeWorld :: ItemName -> Room -> World -> World

Funkcja przyjmuje nazwę przedmiotu oraz pokój i zmienia mapę pokojów poprzez usunięcie przedmiotu z pokoju funkcją takeItemChangeRoom.

takeItemChangeRoom :: ItemName -> Room -> Room

Funkcja przyjmuje nazwę przedmiotu oraz pokój. Zwraca pokój z usuniętym przedmiotem.

takeItemChangeGame :: ItemName -> AdventureGame -> AdventureGame

Funkcja przyjmuje nazwę przedmiotu i zmienia stan gry poprzez usunięcie go z pokoju.

takeItem :: ItemName -> GameIO Bool

Funkcja przyjuje nazwę przedmiotu, sprawda czy jest on w aktualnym pokoju. Jeśli przedmiot jest dostępny stan gry zostaje zmodyfikowany poprzez zabranie przedmiotu z pokoju i umieszczenie go w inwertarzu. W przeciwnym przypadku jest wypisywany odpowiedni komunikat stan gry nie zostaje zmodyfikowany.

putItemChangeGame :: ItemName -> AdventureGame -> AdventureGame

Funkcja przyjmuje nazwę przedmiotu i zmienia stan gry poprzez dodanie go do pokoju.

putItemChangeWorld :: ItemName -> Room -> World -> World

Funkcja przyjmuje nazwę przedmiotu oraz pokój i zmienia mapę pokojów poprzez dodanie przedmiotu z pokoju funkcją putItemChangeRoom.

putItemChangeRoom :: ItemName -> Room -> Room

Funkcja przyjmuje nazwę przedmiotu oraz pokój. Zwraca pokój z dodanym przedmiotem i domyślym opisem lokalizacji.

putItem :: ItemName -> GameIO Bool

Funkcja przyjuje nazwę przedmiotu, sprawda czy jest on w inwentarzu. Jeśli przedmiot jest dostępny stan gry zostaje zmodyfikowany poprzez położenie przedmiotu w pokoju. W przeciwnym przypadku jest wypisywany odpowiedni komunikat stan gry nie zostaje zmodyfikowany.

showInventory :: GameIO Bool

Funkcja wypisuje na konsolę aktualny stan posiadanych przedmiotów.